+++ src/stories/bookShelf/BookShelfCard.stories.tsx
@@ -1,4 +1,4 @@
-import BookShelf from '@/v1/bookShelf/BookShelf';
+import BookShelf from '@/v1/bookShelf/BookShelfCard';
 import { Meta, StoryObj } from '@storybook/react';
 
 const meta: Meta<typeof BookShelf> = {
